GOVSI-719 - Move token validation into a seperate service
- The TokenService is getting heavy with a lot of stuff. Simplify that by moving out the Token validation logic into a TokenValidationService
