+++ oidc-api/src/test/java/uk/gov/di/authentication/oidc/lambda/LogoutHandlerTest.java
@@ -23,7 +23,7 @@ import uk.gov.di.authentication.shared.services.ClientSessionService;
 import uk.gov.di.authentication.shared.services.ConfigurationService;
 import uk.gov.di.authentication.shared.services.DynamoClientService;
 import uk.gov.di.authentication.shared.services.SessionService;
-import uk.gov.di.authentication.shared.services.TokenService;
+import uk.gov.di.authentication.shared.services.TokenValidationService;
 
 import java.net.URI;
 import java.time.LocalDateTime;
@@ -51,7 +51,8 @@ class LogoutHandlerTest {
     private final SessionService sessionService = mock(SessionService.class);
     private final DynamoClientService dynamoClientService = mock(DynamoClientService.class);
     private final ClientSessionService clientSessionService = mock(ClientSessionService.class);
-    private final TokenService tokenService = mock(TokenService.class);
+    private final TokenValidationService tokenValidationService =
+            mock(TokenValidationService.class);
 
     private static final State STATE = new State();
     private static final String COOKIE = "Cookie";
@@ -72,7 +73,7 @@ class LogoutHandlerTest {
                         sessionService,
                         dynamoClientService,
                         clientSessionService,
-                        tokenService);
+                        tokenValidationService);
         when(configurationService.getDefaultLogoutURI()).thenReturn(DEFAULT_LOGOUT_URI);
         ECKey ecSigningKey =
                 new ECKeyGenerator(Curve.P_256).algorithm(JWSAlgorithm.ES256).generate();
@@ -85,7 +86,8 @@ class LogoutHandlerTest {
     public void shouldDeleteSessionAndRedirectToClientLogoutUriForValidLogoutRequest() {
         when(dynamoClientService.getClient("client-id"))
                 .thenReturn(Optional.of(createClientRegistry()));
-        when(tokenService.validateIdTokenSignature(signedIDToken.serialize())).thenReturn(true);
+        when(tokenValidationService.validateIdTokenSignature(signedIDToken.serialize()))
+                .thenReturn(true);
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
         event.setHeaders(Map.of(COOKIE, buildCookieString(CLIENT_SESSION_ID)));
         event.setQueryStringParameters(
@@ -109,7 +111,8 @@ class LogoutHandlerTest {
     public void shouldNotReturnStateWhenStateIsNotSentInRequest() {
         when(dynamoClientService.getClient("client-id"))
                 .thenReturn(Optional.of(createClientRegistry()));
-        when(tokenService.validateIdTokenSignature(signedIDToken.serialize())).thenReturn(true);
+        when(tokenValidationService.validateIdTokenSignature(signedIDToken.serialize()))
+                .thenReturn(true);
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
         event.setHeaders(Map.of(COOKIE, buildCookieString(CLIENT_SESSION_ID)));
         event.setQueryStringParameters(
@@ -204,7 +207,8 @@ class LogoutHandlerTest {
         SignedJWT signedJWT =
                 TokenGeneratorHelper.generateIDToken(
                         "invalid-client-id", new Subject(), "http://localhost-rp", ecSigningKey);
-        when(tokenService.validateIdTokenSignature(signedJWT.serialize())).thenReturn(true);
+        when(tokenValidationService.validateIdTokenSignature(signedJWT.serialize()))
+                .thenReturn(true);
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
         event.setHeaders(Map.of(COOKIE, buildCookieString(CLIENT_SESSION_ID)));
         event.setQueryStringParameters(
@@ -233,7 +237,8 @@ class LogoutHandlerTest {
 
     @Test
     public void shouldRedirectToDefaultLogoutUriWhenLogoutUriInRequestDoesNotMatchClientRegistry() {
-        when(tokenService.validateIdTokenSignature(signedIDToken.serialize())).thenReturn(true);
+        when(tokenValidationService.validateIdTokenSignature(signedIDToken.serialize()))
+                .thenReturn(true);
         when(dynamoClientService.getClient("client-id"))
                 .thenReturn(Optional.of(createClientRegistry()));
         APIGatewayProxyRequestEvent event = new APIGatewayProxyRequestEvent();
